requirements.txt
requests==2.10.0
python-twitter==3.1
